requirements_new.txt
aldryn-apphooks-config==0.6.0
arrow==1.2.2
asgiref==3.5.0
beautifulsoup4==4.10.0
binaryornot==0.4.4
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.0
cookiecutter==1.7.3
cssselect2==0.5.0
dj-database-url==0.5.0
Django==3.2.12
django-appconf==1.0.5
django-appdata==0.3.2
django-browser-reload==1.3.0
django-classy-tags==3.0.1
django-cms==3.10.0
django-colorfield==0.6.3
django-cookie-consent==0.3.1
django-countries==7.3.2
django-filer==2.1.2
django-formtools==2.3
django-js-asset==2.0.0
django-livereload-server==0.4
django-meta==2.0.0
django-mptt==0.13.4
django-parler==2.3
django-polymorphic==3.0.0
django-sekizai==3.0.1
django-sortedm2m==3.1.1
django-taggit==2.1.0
django-taggit-autosuggest==0.3.8
django-taggit-templatetags==0.2.5
django-tailwind==3.1.1
django-templatetag-sugar==1.0
django-treebeard==4.5.1
djangocms-admin-style==3.1.0
djangocms-apphook-setup==0.4.1
djangocms-attributes-field==2.0.0
djangocms-blog==1.2.3
djangocms-file==3.0.0
djangocms-googlemap==2.0.0
djangocms-history==2.0.0
djangocms-icon==2.0.0
djangocms-installer==2.0.0
djangocms-link==3.0.0
djangocms-picture==4.0.0
djangocms-style==3.0.0
djangocms-text-ckeditor==5.0.1
djangocms-video==3.0.0
easy-thumbnails==2.8.1
html5lib==1.1
idna==3.3
Jinja2==3.1.1
jinja2-time==0.2.0
lxml==4.8.0
MarkupSafe==2.1.1
Pillow==9.0.1
poyo==0.5.0
python-dateutil==2.8.2
python-slugify==6.1.1
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
reportlab==3.6.9
requests==2.27.1
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
svglib==1.2.1
text-unidecode==1.3
tinycss2==1.1.1
tornado==6.1
typing_extensions==4.1.1
tzdata==2022.1
tzlocal==4.1
Unidecode==1.3.4
urllib3==1.26.9
webencodings==0.5.1
